requirements.txt
pygame
tensorflow
scipy
pillow
pandas